# Conflicts:
#	Stratis.Bitcoin.Features.Wallet.Tests/WalletControllerTest.cs
#	Stratis.Bitcoin.Features.Wallet.Tests/WalletTransactionHandlerTest.cs
#	Stratis.Bitcoin.Features.Wallet/Controllers/WalletController.cs

+++ Stratis.Bitcoin/BlockPulling/PullerDownloadAssignments.cs
@@ -112,7 +112,7 @@ namespace Stratis.Bitcoin.BlockPulling
                     int peerTaskCount = filteredPeers[i].TasksAssignedCount;
                     if (peerTaskCount > HighWorkAmountThreshold)
                     {
-                        double penaltyCoef = 1 + ((peerTaskCount * peerTaskCount) / (totalWork * totalWork));
+                        double penaltyCoef = 1 + ((double)(peerTaskCount * peerTaskCount) / (totalWork * totalWork));
                         scores[i] /= penaltyCoef;
                     }
                 }
